The _ variable is not meant for binding semnatically, so we can avoid assigning it for better performance.

For `benchmarks/pattern_matching/underscore.rb`:

```
# Without optimization
Comparison:
transpiled (last pattern):   2575858.5 i/s 
  baseline (last pattern):   1625193.4 i/s - 1.58x  slower

# With optimization
Comparison:
transpiled (last pattern):   2212582.0 i/s 
  baseline (last pattern):   1632729.4 i/s - 1.36x  slower
```